+++ AllowAnyNames/_Thunderstore/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2026-Mar-11
+
+### Added
+
+- Added 'ChatDisplayRTFNames' config option that allows specifying whenever to use or not use RTF names as part of chat
+- Added EasySettings as a soft dependency for configuring the option from above
+
 ## [2.0.4] - 2026-Jan-08
 
 ### Fixed